net: emaclite: fix compile warning in BUFFER_ALIGN
Use ulong instead of u32 to fix compile warning when building
kernel for arm64 (zynqmp):

drivers/net/ethernet/xilinx/xilinx_emaclite.c:98:42: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]
   98 | #define BUFFER_ALIGN(adr) ((ALIGNMENT - ((u32)adr)) % ALIGNMENT)
      |                                          ^

+++ drivers/net/ethernet/xilinx/xilinx_emaclite.c
@@ -98,7 +98,7 @@
 #define ALIGNMENT		4
 
 /* BUFFER_ALIGN(adr) calculates the number of bytes to the next alignment. */
-#define BUFFER_ALIGN(adr) ((ALIGNMENT - ((u32)adr)) % ALIGNMENT)
+#define BUFFER_ALIGN(adr) ((ALIGNMENT - ((ulong)adr)) % ALIGNMENT)
 
 #ifdef __BIG_ENDIAN
 #define xemaclite_readl		ioread32be
